01_build_langgraph/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cattrs==24.1.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
curl_cffi==0.11.1
dataclasses-json==0.6.7
distro==1.9.0
fredapi==0.5.2
frozendict==2.4.6
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
ibm-cos-sdk==2.14.1
ibm-cos-sdk-core==2.14.1
ibm-cos-sdk-s3transfer==2.14.1
ibm_watsonx_ai==1.3.20
idna==3.10
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-ibm==0.3.11
langchain-mcp-adapters==0.1.1
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
lomond==0.3.3
lxml==5.4.0
markdown-it-py==3.0.0
marshmallow==3.26.1
mcp==1.9.0
mdurl==0.1.2
multidict==6.4.4
multitasking==0.0.11
mypy_extensions==1.1.0
numpy==2.2.6
openai==1.82.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.31.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyrate-limiter==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.2
requests-cache==1.2.1
requests-ratelimiter==0.7.0
requests-toolbelt==1.0.0
rich==14.0.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.5
starlette==0.46.2
ta_lib_easy==1.0.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
yfinance==0.2.61
zstandard==0.23.0
fastapi
